requirements.txt
aiolimiter==1.1.0
anyio==4.0.0
APScheduler==3.10.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
pyTelegramBotAPI==4.14.0
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
telebot==0.0.5
tornado==6.3.3
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
